make schema and models compatible

+++ src/models.rs
@@ -3,39 +3,38 @@ use diesel::*;
 use diesel_geometry::pg::data_types::PgPoint;
 use chrono::NaiveDateTime;
 use uuid::Uuid;
-
 use crate::schema::*;
 
-#[derive(Insertable, QueryableByName, AsChangeset, Associations, Identifiable, Debug)]
-#[table_name = "weather"]
+#[derive(Insertable, Queryable, AsChangeset, Associations, Identifiable, Debug)]
+#[table_name="weather"]
 pub struct Weather {
     pub id: Uuid,
     pub location: PgPoint,
     pub timestamp: NaiveDateTime,
 }
 
-#[derive(Insertable, QueryableByName, AsChangeset, Associations, Identifiable, Debug)]
+#[derive(Insertable, Queryable, Associations, Identifiable, Debug)]
 #[belongs_to(Weather)]
-#[table_name = "water_levels"]
+#[table_name="water_levels"]
 pub struct WaterLevel {
     pub id: Uuid,
     pub location: PgPoint,
     pub timestamp: NaiveDateTime,
     pub weather_id: Uuid,
-    pub level: f32,
+    pub level: f64,
 }
 
-#[derive(Insertable, QueryableByName, AsChangeset, Identifiable, Debug)]
+#[derive(Insertable, Queryable, AsChangeset, Identifiable, Debug)]
 #[primary_key(key)]
-#[table_name = "config"]
+#[table_name="config"]
 pub struct Config {
     pub key: String,
     pub value: String,
     pub timestamp: NaiveDateTime,
 }
 
-#[derive(Insertable, QueryableByName, AsChangeset, Identifiable, Debug)]
-#[table_name = "webhooks"]
+#[derive(Insertable, Queryable, AsChangeset, Identifiable, Debug)]
+#[table_name="webhooks"]
 pub struct Webhook {
     pub id: Uuid,
     pub url: String,
